+++ gui/button.py
@@ -1,0 +1,43 @@
+from PyQt5.QtCore import Qt, QSize
+from PyQt5.QtWidgets import QToolButton, QSizePolicy
+from PyQt5.QtGui import QPixmap, QIcon
+
+
+class Button(QToolButton):
+    def __init__(self, name):
+        super().__init__()
+
+        # transparent background
+        self.setStyleSheet("QToolButton {border-style: outset; border-width: 0px; margin: 0px; padding: 0px;}")
+        self.setAttribute(Qt.WA_TranslucentBackground)
+
+        #picture change
+        self.name = name
+        self._pressed = False
+        self.updatePic()
+
+        self.pressed.connect(self._whenpressed)
+        self.released.connect(self._whenreleased)
+
+    def resizeEvent(self, e):
+        super().resizeEvent(e)
+        self.setIconSize(self.size())
+
+    # size kludges :/
+    def sizeHint(self):
+        size = super().sizeHint()
+        return QSize(size.width(), size.width())
+
+    def _whenpressed(self):
+        self._pressed = True
+        self.updatePic()
+
+    def _whenreleased(self):
+        self._pressed = False
+        self.updatePic()
+
+    # changes picture of button
+    def updatePic(self, name=None):
+        self.name = name or self.name
+        pixmap = QPixmap(self.name + ("pr" if self._pressed else ""))
+        self.setIcon(QIcon(pixmap))
